Metasploit::Concern.run and shared example
MSP-9895

I forgot to call ActiveSupport.run_load_hooks in some other gems, so to
prevent that in the future, add a shared example
"Metasploit::Concern.run" which will check that
`Metasploit::Concern.run(self)` or the lower-level
`ActiveSupport.run_load_hooks(<name>, self)` is called for the
described_class.  The shared example itself is tested with cucumber
which can be run with `rake features` now.