Do not mark base class as unloadable
MSP-12527

The base class cannot be marked as unloadable if the base class was not
autoloaded by ActiveSupport::Dependencies, but instead by
Kernel.autoload (or require) because the class won't be reloaded then,
which leads the base class being created as a dummy Module by
ActiveSupport::Dependencies.  Base classes that are autoloaded by
ActiveSupport::Dependencies don't need to be marked as unloadable
because ActiveSupport::Dependencies automatically unloads classes it
loaded.